SystemCode/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
gensim==4.1.2
keras==2.6.0
neo4j==5.1.0
nltk==3.6.5
numpy==1.19.5
pandas==1.3.4
scikit-learn==1.0.2
scipy==1.6.1
sentence-transformers==2.2.2
spacy==3.4.0
tensorflow==2.6.0
tensorflow-addons==0.15.0
torch==1.12.0
torchvision==0.13.0